+++ app/src/main/java/com/example/instagram/database/entity/Follow.kt
@@ -23,8 +23,8 @@ import androidx.room.PrimaryKey
     ]
 )
 data class Follow(
-    val owner_id: Int,
-    val follower_id: Int,
+    val owner_id: Long,
+    val follower_id: Long,
     @PrimaryKey(autoGenerate = true)
-    val follow_row_id: Int = 0,
+    val follow_row_id: Long = 0,
 )
