+++ app/src/main/java/com/example/instagram/database/entity/Tag.kt
@@ -23,8 +23,8 @@ import androidx.room.PrimaryKey
 )
 
 data class Tag(
-    val post_id: Int,
-    val profile_id: Int,
+    val post_id: Long,
+    val profile_id: Long,
     @PrimaryKey(autoGenerate = true)
-    val tag_id: Int = 0,
+    val tag_id: Long = 0,
 )
